LPS-98795 Merge into a single extender.
